cattle-weight-detection-app/requirements.txt
streamlit>=1.20.0
tensorflow>=2.10.0,<2.16.0
Pillow>=9.0.0
numpy>=1.21.0
